Remove broken lazy-initialization pattern for MathConsts
Any call to ToEuclidean method would fail unless a previous call to
FromEuclidean is executed, as private _m fields was only properly
initialized when FromEuclidean was called.
The content of MathConsts doesn't seem to be so time-consuming or use so
much memory to require a lazy-initialization pattern, so I directly
instantiated the MathConsts in _m field in the constructor.